moar

+++ src/info/kirklander/draw_thai/DrawThai.java
@@ -17,17 +17,17 @@ public class DrawThai extends Activity {
         // these are the goals that the user must move over in order
         // showing Gor Gai now as a proof of concept.
         // TODO: store this data more better like
-        game.addGoalPoint(75, 300);
-        game.addGoalPoint(75, 250);
-        game.addGoalPoint(75, 200);
-        game.addGoalPoint(100, 175);
-        game.addGoalPoint(75, 150);
-        game.addGoalPoint(125, 125);
-        game.addGoalPoint(175, 150);
-        game.addGoalPoint(175, 150);
-        game.addGoalPoint(175, 200);
-        game.addGoalPoint(175, 250);
-        game.addGoalPoint(175, 300);
+        game.addGoalPoint((float) 75.0, (float) 300.0);
+        game.addGoalPoint((float) 75, (float) 250);
+        game.addGoalPoint((float) 75, (float) 200);
+        game.addGoalPoint((float) 100, (float) 175);
+        game.addGoalPoint((float) 75, (float) 150);
+        game.addGoalPoint((float) 125, (float) 125);
+        game.addGoalPoint((float) 175, (float) 150);
+        game.addGoalPoint((float) 175, (float) 150);
+        game.addGoalPoint((float) 175, (float) 200);
+        game.addGoalPoint((float) 175, (float) 250);
+        game.addGoalPoint((float) 175, (float) 300);
         
         // create view and display it
         final GameView gv = new GameView(this, game);
